agents/agent_MCPRedditAgent/requirements.txt
openai
tiktoken
python-dotenv
urllib3
pydantic
aioconsole

asyncpraw